Refactor Brs_023_027
