+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_023_027/V1/Model/StartCalculationCommandV1.cs
@@ -34,7 +34,7 @@ public sealed record StartCalculationCommandV1
         CalculationInputV1 inputParameter)
             : base(
                 operatingIdentity,
-                orchestrationDescriptionUniqueName: new Brs_023_027_V1(),
+                orchestrationDescriptionUniqueName: Brs_023_027.V1,
                 inputParameter)
     {
     }
